requirements.txt
streamlit==1.32.0
yfinance==0.2.36
pandas==2.2.1
plotly==5.19.0
beautifulsoup4==4.12.3
requests==2.31.0
nltk==3.8.1
ta==0.11.0
python-dotenv==1.0.1
httpx==0.27.0
numpy==1.26.4 
openai==1.1.0
